(BSR)[API] refactor: move atomic and its functions to a dedicated module
